chore: Clean up imports and fixture locations
